Fix various minor problems.
Mmakefile:
	Run the realclean_tests target *before* the realclean_subdirs
	target since the former requires scripts/mmake to exist and
	the latter deletes it.

tests/debugger/save.exp:
	Update this expected output to conform to recent changes.
	(The .exp2 file was updated as part of those changes, but
	this one wasn't.)

trace/mercury_trace_cmd_breakpoint.c:
trace/mercury_trace_source.c:
	Avoid warnings from GCC.

+++ trace/mercury_trace_source.c
@@ -264,7 +264,7 @@ MR_trace_source_open_server(MR_TraceSourceServer *server,
     name = MR_malloc(base_len + 10);
     do {
         i++;
-        sprintf(name, "%s.%d", MR_SOURCE_SERVER_BASENAME, i);
+        sprintf(name, "%s.%lu", MR_SOURCE_SERVER_BASENAME, (unsigned long)i);
         /*
         ** This should fail if there is no server with this
         ** name.
